requirements.txt
pyserial
paho-mqtt
